Refactor currently WIP models into broken_models folder.
