+++ guidance/models/vertexai/__init__.py
@@ -1,3 +1,3 @@
 from ._PaLM2 import PaLM2Completion, PaLM2Chat, PaLM2Instruct
 from ._Codey import CodeyCompletion, CodeyInstruct, CodeyChat
-from ._Gemini import GeminiChat
+from ..broken_models._Gemini import GeminiChat
